+++ packages/carp_context_package/lib/src/air_quality/air_quality_measure.dart
@@ -17,11 +17,11 @@ class AirQualityMeasure extends Measure {
   AirQualityMeasure(MeasureType type, {name, enabled, this.apiKey})
       : super(type, name: name, enabled: enabled);
 
-  static Function get fromJsonFunction => _$AirQualityMeasureFromJson;
+  Function get fromJsonFunction => _$AirQualityMeasureFromJson;
 
   factory AirQualityMeasure.fromJson(Map<String, dynamic> json) =>
-      FromJsonFactory.fromJson(
-          json[Serializable.CLASS_IDENTIFIER].toString(), json);
+      FromJsonFactory()
+          .fromJson(json[Serializable.CLASS_IDENTIFIER].toString(), json);
 
   Map<String, dynamic> toJson() => _$AirQualityMeasureToJson(this);