+++ packages/carp_context_package/lib/src/context_package.dart
@@ -55,20 +55,28 @@ class ContextSamplingPackage implements SamplingPackage {
   }
 
   void onRegister() {
-    FromJsonFactory.registerFromJsonFunction("LocationMeasure", LocationMeasure.fromJsonFunction);
-    FromJsonFactory.registerFromJsonFunction("WeatherMeasure", WeatherMeasure.fromJsonFunction);
-    FromJsonFactory.registerFromJsonFunction("GeofenceMeasure", GeofenceMeasure.fromJsonFunction);
-    FromJsonFactory.registerFromJsonFunction("AirQualityMeasure", AirQualityMeasure.fromJsonFunction);
-    FromJsonFactory.registerFromJsonFunction("GeoPosition", GeoPosition.fromJsonFunction);
-    FromJsonFactory.registerFromJsonFunction("MobilityMeasure", MobilityMeasure.fromJsonFunction);
+    FromJsonFactory().register(LocationMeasure(null));
+    FromJsonFactory().register(WeatherMeasure(null));
+    FromJsonFactory().register(GeofenceMeasure(null));
+    FromJsonFactory().register(AirQualityMeasure(null));
+    FromJsonFactory().register(GeoPosition(null));
+    FromJsonFactory().register(MobilityMeasure(null));
 
     // registering the transformers from CARP to OMH for geolocation and physical activity.
     // we assume that there is an OMH schema registered already...
-    TransformerSchemaRegistry().lookup(NameSpace.OMH).add(LOCATION, OMHGeopositionDatum.transformer);
-    TransformerSchemaRegistry().lookup(NameSpace.OMH).add(ACTIVITY, OMHPhysicalActivityDatum.transformer);
+    TransformerSchemaRegistry()
+        .lookup(NameSpace.OMH)
+        .add(LOCATION, OMHGeopositionDatum.transformer);
+    TransformerSchemaRegistry()
+        .lookup(NameSpace.OMH)
+        .add(ACTIVITY, OMHPhysicalActivityDatum.transformer);
   }
 
-  List<Permission> get permissions => [Permission.locationAlways, Permission.sensors, Permission.activityRecognition];
+  List<Permission> get permissions => [
+        Permission.locationAlways,
+        Permission.sensors,
+        Permission.activityRecognition
+      ];
 
   SamplingSchema get common => SamplingSchema()
     ..type = SamplingSchemaType.COMMON
@@ -77,7 +85,8 @@ class ContextSamplingPackage implements SamplingPackage {
     ..measures.addEntries([
       MapEntry(
         LOCATION,
-        Measure(MeasureType(NameSpace.CARP, LOCATION), name: 'Location', enabled: true),
+        Measure(MeasureType(NameSpace.CARP, LOCATION),
+            name: 'Location', enabled: true),
       ),
       MapEntry(
           GEOLOCATION,
@@ -91,20 +100,28 @@ class ContextSamplingPackage implements SamplingPackage {
           )),
       MapEntry(
         ACTIVITY,
-        Measure(MeasureType(NameSpace.CARP, ACTIVITY), name: 'Activity Recognition', enabled: true),
+        Measure(MeasureType(NameSpace.CARP, ACTIVITY),
+            name: 'Activity Recognition', enabled: true),
       ),
       MapEntry(
           WEATHER,
           WeatherMeasure(MeasureType(NameSpace.CARP, WEATHER),
-              name: 'Local Weather', enabled: true, apiKey: '12b6e28582eb9298577c734a31ba9f4f')),
+              name: 'Local Weather',
+              enabled: true,
+              apiKey: '12b6e28582eb9298577c734a31ba9f4f')),
       MapEntry(
           AIR_QUALITY,
           AirQualityMeasure(MeasureType(NameSpace.CARP, AIR_QUALITY),
-              name: 'Local Air Quality', enabled: true, apiKey: '9e538456b2b85c92647d8b65090e29f957638c77')),
+              name: 'Local Air Quality',
+              enabled: true,
+              apiKey: '9e538456b2b85c92647d8b65090e29f957638c77')),
       MapEntry(
           GEOFENCE,
           GeofenceMeasure(MeasureType(NameSpace.CARP, GEOFENCE),
-              enabled: true, center: GeoPosition(55.7943601, 12.4461956), radius: 500, name: 'Geofence (Virum)')),
+              enabled: true,
+              center: GeoPosition(55.7943601, 12.4461956),
+              radius: 500,
+              name: 'Geofence (Virum)')),
       MapEntry(
           MOBILITY,
           MobilityMeasure(MeasureType(NameSpace.CARP, MOBILITY),