+++ packages/carp_survey_package/lib/survey.g.dart
@@ -8,10 +8,11 @@ part of survey;
 
 RPTaskMeasure _$RPTaskMeasureFromJson(Map<String, dynamic> json) {
   return RPTaskMeasure(
-    json['type'] == null
+    type: json['type'] == null
         ? null
         : MeasureType.fromJson(json['type'] as Map<String, dynamic>),
     name: json['name'] as String,
+    description: json['description'] as String,
     enabled: json['enabled'] as bool,
   )
     ..$type = json[r'$type'] as String
@@ -32,6 +33,7 @@ Map<String, dynamic> _$RPTaskMeasureToJson(RPTaskMeasure instance) {
   writeNotNull(r'$type', instance.$type);
   writeNotNull('type', instance.type);
   writeNotNull('name', instance.name);
+  writeNotNull('description', instance.description);
   writeNotNull('enabled', instance.enabled);
   writeNotNull('configuration', instance.configuration);
   return val;